fix(knowledge): fix developer permission check and organization namespace handling
- Fix developer restriction check to use model_fields_set instead of getattr
  This ensures explicit null values are properly blocked for restricted fields
  (retrieval_config, summary_enabled, summary_model_ref, etc.)

- Fix organization namespace check order in _get_user_kb_permission_level
  Move organization check before group check to ensure correct permission logic

- Add organization admin guard to update_document_content method
  Only admin users can edit documents in organization knowledge bases

- Fix indentation error in update_document_content method

All 49 tests pass including 32 group_kb permission tests
